Tracker now uses Exporter which saves positions every 15 mn / and in live logs

+++ tests/src/exporter.js
@@ -2,6 +2,7 @@ const fs = require('fs');
 const os = require('os');
 
 const expect = require('chai').expect;
+const Promise = require('bluebird');
 
 const Exporter = require('../../src/exporter');
 
@@ -10,20 +11,29 @@ const config = require('../../config');
 describe('exporter', () => {
   const tmpdir = os.tmpdir();
   const currentYYYYMMDD = new Date().toISOString().slice(0, 10).replace(/-/g, "");
-  const todayLog = `/tmp/aabbccddeeff-${currentYYYYMMDD}.json`;
+  const todayLog = `${tmpdir}/aabbccddeeff-${currentYYYYMMDD}.json`;
+  const liveLog = `${tmpdir}/aabbccddeeff.json`;
+  let exporter = null;
 
   before(() => {
     config.dashboard.base = '/tmp/';
+    config.dashboard.autosaveInterval = 90;
   });
-  afterEach(() => {
-    if (fs.existsSync(todayLog)) {
-      fs.unlinkSync(todayLog);
-    }
+  afterEach(async() => {
+    await exporter.close();
+    if (fs.existsSync(todayLog)) fs.unlinkSync(todayLog);
+    if (fs.existsSync(liveLog)) fs.unlinkSync(liveLog);
+  });
+  after(() => {
+    config.dashboard.autosaveInterval = 300 * 1000;
   });
 
-  it('should create and append to today log', async() => {
-    const exporter = new Exporter('aabbccddeeff', tmpdir);
+  it('should create and append to today log in live logs and in resident logs', async() => {
+    exporter = new Exporter('aabbccddeeff', tmpdir);
     await exporter.append({ pi1: -62, pi2: -62, pi3: -78 }, { x: 5, y: 2 });
+
+    expect(fs.existsSync(todayLog)).to.equal(false);
+    expect(fs.existsSync(`${tmpdir}/aabbccddeeff.json`)).to.equal(true);
     await exporter.saveCurrent();
 
     expect(fs.existsSync(todayLog)).to.be.true;
@@ -34,11 +44,20 @@ describe('exporter', () => {
     fs.writeFileSync(todayLog, '{"data":[{"pool":{"pi1":-62,"pi2":-62,"pi3":-78},"coordinates":{"x":5,"y":2}}]}');
     expect(fs.readFileSync(todayLog, 'utf8')).to.equal('{"data":[{"pool":{"pi1":-62,"pi2":-62,"pi3":-78},"coordinates":{"x":5,"y":2}}]}');
 
-    const exporter = new Exporter('aabbccddeeff', tmpdir);
+    exporter = new Exporter('aabbccddeeff', tmpdir);
     await exporter.append({ pi1: -90 }, null);
     await exporter.saveCurrent();
 
     expect(fs.existsSync(todayLog)).to.be.true;
     expect(fs.readFileSync(todayLog, 'utf8')).to.equal('{"data":[{"pool":{"pi1":-62,"pi2":-62,"pi3":-78},"coordinates":{"x":5,"y":2}},{"pool":{"pi1":-90},"coordinates":null}]}');
+  });
+
+  it('should autosave today current logs', async() => {
+    exporter = new Exporter('aabbccddeeff', tmpdir);
+    await exporter.append({ pi1: -62 });
+
+    expect(fs.existsSync(todayLog)).to.be.false;
+    await Promise.delay(200);
+    expect(fs.existsSync(todayLog)).to.be.true;
   });
 });
